Make cliqueEpochTransitionSigners a standalone function
